Refactor xtransport
